requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.5
django-jazzmin==3.0.1
logger==1.4
lxml==6.0.0
mysqlclient==2.2.7
python-decouple==3.8
python-docx==1.2.0
sqlparse==0.5.3
typing_extensions==4.14.1
